requirements.txt
Flask==2.1.1
gevent==21.12.0
flasgger==0.9.5
Flask_SQLAlchemy==2.5.1
SQLAlchemy==1.4.35
psycopg2-binary==2.9.3
python-dotenv==0.19.2
flask_marshmallow==0.14.0
marshmallow_sqlalchemy==0.28.0
Flask-UUID==0.2
Flask-RESTful==0.3.9
marshmallow==3.15.0
Flask-JWT-Extended==4.3.1
redis==4.2.2
pydantic==1.9.0
backoff==1.11.1
Flask-Migrate==3.1.0
alembic==1.7.7
Authlib==1.0.1
requests==2.27.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-instrumentation-requests==0.30b1
